backend/admin-cleanup/requirements.txt
psycopg2-binary==2.9.9
google-auth==2.23.0
google-api-python-client==2.100.0